backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
pydantic-settings==2.1.0
pyyaml
pandas
psycopg2-binary
SQLAlchemy==1.4.49
requests